move peer database-related methods to a new module
