+++ src/peers/mod.rs
@@ -1,3 +1,4 @@
+pub mod database;
 pub mod discovery;
 mod hello;
 pub mod local_ips;
